BayesPois(Bolstad2)	R Documentation

Bayesian Pois Regression

Description

Performs Metropolis Hastings on the logistic regression model to draw sample from posterior. Uses a matched curvature Student's t candidate generating distribution with 4 degrees of freedom to give heavy tails.

Usage

BayesPois(y, x, steps = 1000,
                 priorMean = NULL, priorVar = NULL,
                 mleMean = NULL, mleVar,
                 startValue = NULL, randomSeed = NULL,
                 plots = FALSE)


Arguments

y 	 the binary response vector

x 	 matrix of covariates

steps 	 the number of steps to use in the Metropolis-Hastings updating

priorMean 	 the mean of the prior

priorVar 	 the variance of the prior

mleMean 	 the mean of the matched curvature likelihood

mleVar 	 the covariance matrix of the matched curvature likelihood

startValue

randomSeed 	 a random seed to use for different chains

plots 	 Plot the time series and auto correlation functions for each of the model coefficients

Value

A list containing the following components:

beta 	 a data frame containing the sample of the model coefficients from the posterior distribution

mleMean 	 the mean of the matched curvature likelihood. This is useful if you've used a training set to estimate the value and wish to use it with another data set

mleVar 	 the covariance matrix of the matched curvature likelihood. See mleMean for why you'd want this

Examples


data(poissonTest.df)
results <- BayesPois(poissonTest.df$y, poissonTest.df$x)

[Package Bolstad2 version 1.0-14 Index]