credIntSamp(bolstad2)	R Documentation

Calculate a credible interval from a numerically specified posterior CDF

Description

Calculates a lower, upper, or two-sided credible interval from the numerical posterior CDF.

Usage

credIntSamp(theta, conf = 0.95, type="twosided")

Arguments

theta 	 a sample from the posterior density

conf 	 the desired 'confidence' level

type 	 the type of interval to return, 'lower' = one sided lower bound, 'two-sided' = two - sided, or 'upper' = one sided upper bound. It is sufficient to use 'l','t' or 'u'

Details

This function uses linear interpolation to calculate bounds for points that may not be specified by CDF

Value

a list containing the elements lower.bound, uppper.bound or both depending on type

Examples


## posterior is N(0,1)
theta <- rnorm(1000)
ci<-credIntSamp(theta)
plot(density(theta))
abline(v=c(unlist(ci)))

[Package bolstad2 version 1.0-0 Index]